feat(chat): Channel부분을 Chat으로 바꾸기 쓸모없는 소스 삭제
