+++ backend/src/chat/chat.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
-import { ChannelService } from './channel.service';
+import { ChatService } from './chat.service';
 import { Channel } from 'src/type/chatType';
 import { Gateway2faGuard } from 'src/guard/gateway2fa.guard';
 import { UseGuards } from '@nestjs/common';
@@ -8,12 +8,12 @@ import { UseGuards } from '@nestjs/common';
 @Controller('channel')
 @ApiTags('channel')
 @UseGuards()
-export class ChannelController {
-  constructor(private readonly channelService: ChannelService) {
+export class ChatController {
+  constructor(private readonly chatService: ChatService) {
   }
 
   @Get('public-list')
   getPublicChannels(): Channel[] {
-    return this.channelService.getPublicChannels();
+    return this.chatService.getPublicChannels();
   }
 }
