+++ backend/src/chat/chat.module.ts
@@ -1,8 +1,22 @@
 import { Module } from '@nestjs/common';
 import { ChatGateway } from './chat.gateway';
 import { ChatService } from './chat.service';
+import { ChatController } from './chat.controller';
+import { JwtModule } from '@nestjs/jwt';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { User } from 'src/entity/User';
+import { UserModule } from 'src/user.module';
 
 @Module({
-  providers: [ChatGateway, ChatService]
+  imports: [
+    TypeOrmModule.forFeature([ User ]),
+    JwtModule.register({
+      global: true,
+      secret: process.env.JWT_SECRET,
+      signOptions: { expiresIn: '1d'}
+    }),
+    UserModule],
+  providers: [ChatGateway, ChatService],
+  controllers: [ChatController]
 })
 export class ChatModule {}
